requirements.txt
fastapi
uvicorn
openai
python-jose[cryptography]
passlib[bcrypt]
bcrypt<4.0.0
python-multipart
sqlalchemy
email-validator
pymysql
python-dotenv
